@@ -35,14 +35,14 @@ def apply_game_patches(editor: PatcherEditor, configuration: dict):
 def _modify_raven_beak_damage_table(editor: PatcherEditor, mode: str):
     rb_bmsad = editor.get_file("actors/characters/chozocommander/charclasses/chozocommander.bmsad", Bmsad)
 
-    life_component = rb_bmsad.raw.property.components.LIFE
-    ai_component = rb_bmsad.raw.property.components.AI
+    life_component = rb_bmsad.components["LIFE"]
+    ai_component = rb_bmsad.components["AI"]
 
     if mode == "consistent_high":
-        base_factor = life_component.fields.fields.oDamageSourceFactor # Base damage during regular fight
+        base_factor = life_component.fields.oDamageSourceFactor  # Base damage during regular fight
         counter_factors = [
-            ai_component.fields.fields.oDamageSourceFactorShortShootingGrab, # Most counter cutscenes
-            ai_component.fields.fields.oDamageSourceFactorLongShootingGrab, # Unknown alternate counter cutscene
+            ai_component.fields.oDamageSourceFactorShortShootingGrab,  # Most counter cutscenes
+            ai_component.fields.oDamageSourceFactorLongShootingGrab,  # Unknown alternate counter cutscene
         ]
 
         # Buffs Wave Beam and Ice Missiles to have the same damage VALUES (not factors) as vanilla Plasma Beam and
@@ -68,9 +68,9 @@ def _modify_raven_beak_damage_table(editor: PatcherEditor, mode: str):
         # Debuffs all weapons prior to Wave Beam and Ice Missiles using the same damage factor as Wave Beam and
         # Ice Missiles have in vanilla
         factors = [
-            life_component.fields.fields.oDamageSourceFactor, # Base damage during regular fight
-            ai_component.fields.fields.oDamageSourceFactorShortShootingGrab, # Most counter cutscenes
-            ai_component.fields.fields.oDamageSourceFactorLongShootingGrab, # Unknown alternate counter cutscene
+            life_component.fields.oDamageSourceFactor,  # Base damage during regular fight
+            ai_component.fields.oDamageSourceFactorShortShootingGrab,  # Most counter cutscenes
+            ai_component.fields.oDamageSourceFactorLongShootingGrab,  # Unknown alternate counter cutscene
         ]
 
         for factor in factors:
@@ -103,17 +103,18 @@ def _remove_grapple_blocks(editor: PatcherEditor, configuration: dict):
 def _remove_pb_weaknesses(editor: PatcherEditor):
     # enky
     warlotus = editor.get_file("actors/characters/warlotus/charclasses/warlotus.bmsad", Bmsad)
-    warlotus.raw.property.components.LIFE.fields.fields.bShouldDieWithPowerBomb = False
-    warlotus.raw.property.components.LIFE.fields.fields.oDamageSourceFactor.fPowerBombFactor = 0.0
+    warlotus.components["LIFE"].fields.bShouldDieWithPowerBomb = False
+    warlotus.components["LIFE"].fields.oDamageSourceFactor.fPowerBombFactor = 0.0
 
     # charge door
     for door in ["doorchargecharge", "doorchargeclosed", "doorclosedcharge"]:
         charge_door = editor.get_file(f"actors/props/{door}/charclasses/{door}.bmsad", Bmsad)
-        func = charge_door.raw.property.components.LIFE.functions[0]
-        if func.params.Param1.value:
-            func.params.Param1.value = "CHARGE_BEAM"
-        if func.params.Param2.value:
-            func.params.Param2.value = "CHARGE_BEAM"
+        func = charge_door.components["LIFE"].functions[0]
+
+        if func.get_param(1):
+            func.set_param(1, "CHARGE_BEAM")
+        if func.get_param(2):
+            func.set_param(2, "CHARGE_BEAM")
 
 
 def _warp_to_start(editor: PatcherEditor):
